+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * added **ThumbnailGroup** and **ThumbnailItem**
 * added **EventLayout** element.
 * Bug fixed in CheckTree
+* GoogleMap component added.
 
 ## 1.1.5 
 * was update RadioInput css.